Refactor graph. Isolate transition probability computation functionality out of Graph objects to RWGraph objects, and only use Graph for read/write and object conversion. Rename node2vec to pecanpy for clarity: three modes of PecanPy executions.
