+++ src/pecanpy/rw/__init__.py
@@ -1,0 +1,4 @@
+from .dense_rw import DenseRWGraph
+from .sparse_rw import SparseRWGraph
+
+__all__ = ["DenseRWGraph", "SparseRWGraph"]
